+++ test/Object/Code25Test.php
@@ -1,18 +1,17 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-barcode for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-barcode/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-barcode/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Barcode\Object;
+namespace LaminasTest\Barcode\Object;
 
-use Zend\Barcode;
+use Laminas\Barcode;
 
 /**
- * @group      Zend_Barcode
+ * @group      Laminas_Barcode
  */
 class Code25Test extends TestCommon
 {
@@ -83,7 +82,7 @@ class Code25Test extends TestCommon
 
     public function testBadTextDetectedIfChecksumWished()
     {
-        $this->setExpectedException('\Zend\Barcode\Object\Exception\ExceptionInterface');
+        $this->setExpectedException('\Laminas\Barcode\Object\Exception\ExceptionInterface');
         $this->object->setText('a');
         $this->object->setWithChecksum(true);
         $this->object->getText();
@@ -97,7 +96,7 @@ class Code25Test extends TestCommon
 
     public function testCheckParamsWithLowRatio()
     {
-        $this->setExpectedException('\Zend\Barcode\Object\Exception\ExceptionInterface');
+        $this->setExpectedException('\Laminas\Barcode\Object\Exception\ExceptionInterface');
         $this->object->setText('0123456789');
         $this->object->setBarThinWidth(21);
         $this->object->setBarThickWidth(40);
@@ -106,7 +105,7 @@ class Code25Test extends TestCommon
 
     public function testCheckParamsWithHighRatio()
     {
-        $this->setExpectedException('\Zend\Barcode\Object\Exception\ExceptionInterface');
+        $this->setExpectedException('\Laminas\Barcode\Object\Exception\ExceptionInterface');
         $this->object->setText('0123456789');
         $this->object->setBarThinWidth(20);
         $this->object->setBarThickWidth(61);